# docs/walker/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hermes Walker API",
    "version": "1.0.0",
    "description": "On-demand renewal data and CRM writes for the RSG custom GPT."
  },
  "servers": [
    {
      "url": "https://hermes-gretch-u69864.vm.elestio.app",
      "description": "Hermes VPS (production)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "QueueResponse": {
        "type": "object",
        "properties": {
          "data_as_of": {
            "type": "string"
          },
          "days_window": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            }
          }
        }
      },
      "QueueItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "policy_number": {
            "type": "string"
          },
          "line": {
            "type": "string",
            "nullable": true
          },
          "carrier": {
            "type": "string",
            "nullable": true
          },
          "premium": {
            "type": "number",
            "nullable": true
          },
          "days_out": {
            "type": "integer",
            "nullable": true
          },
          "segment": {
            "type": "string",
            "nullable": true
          },
          "risk": {
            "type": "string"
          },
          "owner": {
            "type": "string",
            "nullable": true
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "last_touch": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "data_as_of": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            }
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "policy_number": {
            "type": "string"
          },
          "expiration_date": {
            "type": "string",
            "nullable": true
          },
          "premium": {
            "type": "number",
            "nullable": true
          },
          "risk_status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "QuietLapseResponse": {
        "type": "object",
        "properties": {
          "data_as_of": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LapseItem"
            }
          }
        }
      },
      "LapseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "policy_number": {
            "type": "string"
          },
          "expired": {
            "type": "string",
            "nullable": true
          },
          "premium": {
            "type": "number",
            "nullable": true
          },
          "risk_status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ScoreboardResponse": {
        "type": "object",
        "properties": {
          "data_as_of": {
            "type": "string"
          },
          "total_renewals": {
            "type": "integer"
          },
          "retention_pct": {
            "type": "number",
            "nullable": true
          },
          "renewed_premium": {
            "type": "number"
          },
          "lost_premium": {
            "type": "number"
          },
          "active_premium": {
            "type": "number"
          }
        }
      },
      "RenewalDetailResponse": {
        "type": "object",
        "properties": {
          "data_as_of": {
            "type": "string"
          },
          "renewal_id": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "renewal": {
            "type": "object",
            "properties": {}
          },
          "nowcerts": {
            "type": "object",
            "properties": {}
          },
          "opportunity": {
            "type": "object",
            "properties": {}
          }
        }
      },
      "TouchRequest": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string",
            "default": "lamar"
          },
          "channel": {
            "type": "string",
            "default": "manual"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "WorksheetRequest": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "fields"
        ]
      },
      "FlagRequest": {
        "type": "object",
        "properties": {
          "flag": {
            "type": "string"
          }
        },
        "required": [
          "flag"
        ]
      },
      "HandoffRequest": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          }
        },
        "required": [
          "note"
        ]
      },
      "OutcomeRequest": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "renewed",
              "rewritten",
              "lost_price",
              "lost_coverage",
              "lost_no_response",
              "do_not_renew"
            ]
          },
          "stage": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "decision"
        ]
      },
      "WriteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "touch_logged": {
            "type": "string"
          },
          "updated_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "decision": {
            "type": "string"
          },
          "stage": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
      "HandoffItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "owner": {
            "type": "string",
            "nullable": true
          },
          "stage": {
            "type": "string",
            "nullable": true
          },
          "handoff_notes": {
            "type": "string"
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "decision": {
            "type": "string",
            "nullable": true
          },
          "last_contact": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "HandoffsResponse": {
        "type": "object",
        "properties": {
          "data_as_of": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HandoffItem"
            }
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "data_as_of": {
            "type": "string"
          },
          "renewal_id": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "stage": {
            "type": "string",
            "nullable": true
          },
          "owner": {
            "type": "string",
            "nullable": true
          },
          "days_out": {
            "type": "integer",
            "nullable": true
          },
          "decision": {
            "type": "string",
            "nullable": true
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "handoff_notes": {
            "type": "string",
            "nullable": true
          },
          "last_touch": {
            "type": "object",
            "nullable": true
          },
          "last_contact_date": {
            "type": "string",
            "nullable": true
          },
          "day1_sent_at": {
            "type": "string",
            "nullable": true
          },
          "touch_count": {
            "type": "integer"
          },
          "nowcerts_policies": {
            "type": "integer"
          },
          "next_action": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/walker/queue": {
      "get": {
        "summary": "Get renewal queue",
        "description": "Renewals inside N days, classified at request time. Every response stamped with data_as_of freshness.",
        "operationId": "getQueue",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 60
            },
            "description": "Days-out window (default 60)"
          }
        ],
        "responses": {
          "200": {
            "description": "Queue with classified items",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/renewal/{renewal_id}": {
      "get": {
        "summary": "Get single renewal detail (live from NowCerts)",
        "description": "Single-client truth pulled LIVE from NowCerts through Hermes, plus post-dedup account total, touch history, worksheet state.",
        "operationId": "getRenewalDetail",
        "parameters": [
          {
            "name": "renewal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single-client truth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenewalDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Renewal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/search": {
      "get": {
        "summary": "Search renewals by name or policy number",
        "description": "Find by name or policy number. Handles name variants via ilike.",
        "operationId": "searchRenewals",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Client name or policy number"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/quiet-lapse": {
      "get": {
        "summary": "Expired terms with no successor \u2014 silent churn catcher",
        "operationId": "getQuietLapse",
        "responses": {
          "200": {
            "description": "Lapsed renewals with no successor term",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuietLapseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/scoreboard": {
      "get": {
        "summary": "Retention scoreboard \u2014 retention pct, renewed/lost premium",
        "operationId": "getScoreboard",
        "responses": {
          "200": {
            "description": "Retention metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreboardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/touch/{id}": {
      "post": {
        "summary": "WRITE: Log a touch",
        "description": "WRITE: Log a touch on the Opportunity. This modifies CRM data.",
        "operationId": "postTouch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id from getQueue or searchRenewals results (renewal ID or Opportunity ID)"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TouchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Touch logged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/worksheet/{id}": {
      "patch": {
        "summary": "WRITE: Update worksheet",
        "description": "WRITE: Update worksheet fields on the Opportunity. This modifies CRM data.",
        "operationId": "patchWorksheet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id from getQueue or searchRenewals results (renewal ID or Opportunity ID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorksheetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Worksheet updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/flag/{id}": {
      "post": {
        "summary": "WRITE: Add a complexity flag",
        "description": "WRITE: Add a complexity flag to the Opportunity. May change the renewal owner if the flag implies escalation or delegation. This modifies CRM data.",
        "operationId": "postFlag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id from getQueue or searchRenewals results (renewal ID or Opportunity ID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flag added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/handoff/{id}": {
      "post": {
        "summary": "WRITE: Set handoff notes",
        "description": "WRITE: Set handoff notes on the Opportunity for Lamar or Gretchen. This modifies CRM data.",
        "operationId": "postHandoff",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id from getQueue or searchRenewals results (renewal ID or Opportunity ID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Handoff note set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/outcome/{id}": {
      "post": {
        "summary": "WRITE: Set renewal decision",
        "description": "WRITE: Set the renewal decision and pipeline stage on the Opportunity. This modifies CRM data.",
        "operationId": "postOutcome",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id from getQueue or searchRenewals results (renewal ID or Opportunity ID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Outcome set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/handoffs": {
      "get": {
        "summary": "Get handoff queue",
        "description": "Returns Opportunities with handoff notes. Optionally filter by owner (e.g. owner=Lamar).",
        "operationId": "getHandoffs",
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by renewal owner (Lamar or Gretchen)"
          }
        ],
        "responses": {
          "200": {
            "description": "Handoff queue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/walker/status/{renewal_id}": {
      "get": {
        "summary": "Get renewal status and next action",
        "description": "Synthesized status for a single renewal: current stage, days out, last touch, flags, handoff notes, and a recommended next action based on the follow-up ladder.",
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "renewal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id from getQueue or searchRenewals results (renewal ID or Opportunity ID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Renewal status with next action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}